tests/wpt/tests/navigation-api/navigate-event/navigate-window-open.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="iframe" name="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  window.onload = t.step_func(() => {
    navigation.onnavigate = t.step_func_done(() => {
      assert_unreached("onnavigate should not have fired in source window");
    });
    iframe.contentWindow.navigation.onnavigate = t.step_func_done(e => {
      assert_equals(e.navigationType, "push");
      assert_true(e.cancelable);
      assert_true(e.canIntercept);
      assert_false(e.userInitiated);
      assert_true(e.hashChange);
      assert_equals(e.downloadRequest, null);
      assert_equals(new URL(e.destination.url).hash, "#1");
      assert_true(e.destination.sameDocument);
      assert_equals(e.destination.key, "");
      assert_equals(e.destination.id, "");
      assert_equals(e.destination.index, -1);
      assert_equals(e.formData, null);
      assert_equals(e.sourceElement, null);
      e.preventDefault();
    });

    window.open("/common/blank.html#1", "i");
  });
}, "window.open() fires navigate event in target window but not source");
</script>
