LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-window-open-expected.txt



FAIL window.open() fires navigate event in target window but not source assert_equals: expected "push" but got "replace"
